+++ lib/src/utils/card_parser_util.dart
@@ -1,4 +1,5 @@
 import 'dart:core';
+
 import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
 import 'package:ml_card_scanner/src/model/card_info.dart';
 import 'package:ml_card_scanner/src/utils/string_extension.dart';
@@ -17,11 +18,7 @@ class CardParserUtil {
   Future<CardInfo?> detectCardContent(InputImage inputImage) async {
     var input = await _textDetector.processImage(inputImage);
 
-    var clearElements = input.blocks
-        .map(
-          (e) => e.text.clean(),
-        )
-        .toList();
+    var clearElements = input.blocks.map((e) => e.text.clean()).toList();
 
     try {
       var possibleCardNumber = clearElements.firstWhere((input) {