+++ lib/src/widget/camera_overlay_widget.dart
@@ -7,52 +7,106 @@ class CameraOverlayWidget extends StatelessWidget {
   final Color overlayColorFilter;
 
   const CameraOverlayWidget({
-    super.key,
     required this.cardOrientation,
     required this.overlayBorderRadius,
     required this.overlayColorFilter,
+    super.key,
   });
 
   @override
   Widget build(BuildContext context) {
-    return ColorFiltered(
-      colorFilter: ColorFilter.mode(overlayColorFilter, BlendMode.srcOut),
-      child: Stack(
-        children: [
-          Container(
-            decoration: const BoxDecoration(
-              color: Colors.transparent,
-            ),
-            child: Align(
-              alignment: Alignment.bottomRight,
-              child: Center(
-                child: _getContainer(context),
-              ),
-            ),
-          ),
-        ],
+    final size = MediaQuery.sizeOf(context);
+
+    return Container(
+      decoration: ShapeDecoration(
+        shape: _ScannerOverlayShape(
+          cutOutHeight: cardOrientation == CardOrientation.portrait
+              ? (size.width * 0.75) * 1.6
+              : (size.width * 0.95) / 1.6,
+          cutOutWidth: cardOrientation == CardOrientation.portrait
+              ? size.width * 0.75
+              : size.width * 0.95,
+          overlayColor: overlayColorFilter,
+          radius: const Radius.circular(25),
+        ),
       ),
     );
   }
+}
 
-  Widget _getContainer(context) {
-    if (cardOrientation == CardOrientation.portrait) {
-      return Container(
-        height: (MediaQuery.of(context).size.width * 0.75) * 1.6,
-        width: (MediaQuery.of(context).size.width * 0.75),
-        decoration: BoxDecoration(
-          color: Colors.black,
-          borderRadius: BorderRadius.circular(overlayBorderRadius),
-        ),
-      );
+class _ScannerOverlayShape extends ShapeBorder {
+  const _ScannerOverlayShape({
+    required this.cutOutWidth,
+    required this.cutOutHeight,
+    required this.overlayColor,
+    required this.radius,
+  });
+
+  final Color overlayColor;
+  final double cutOutWidth;
+  final double cutOutHeight;
+  final Radius radius;
+
+  @override
+  EdgeInsetsGeometry get dimensions => EdgeInsets.zero;
+
+  @override
+  Path getInnerPath(Rect rect, {TextDirection? textDirection}) {
+    return Path()
+      ..fillType = PathFillType.evenOdd
+      ..addPath(getOuterPath(rect), Offset.zero);
+  }
+
+  @override
+  Path getOuterPath(Rect rect, {TextDirection? textDirection}) {
+    Path getLeftTopPath(Rect rect) {
+      return Path()
+        ..moveTo(rect.left, rect.bottom)
+        ..lineTo(rect.left, rect.top)
+        ..lineTo(rect.right, rect.top);
     }
-    return Container(
-      height: (MediaQuery.of(context).size.width * 0.95) / 1.6,
-      width: (MediaQuery.of(context).size.width * 0.95),
-      decoration: BoxDecoration(
-        color: Colors.black,
-        borderRadius: BorderRadius.circular(overlayBorderRadius),
-      ),
+
+    return getLeftTopPath(rect)
+      ..lineTo(rect.right, rect.bottom)
+      ..lineTo(rect.left, rect.bottom)
+      ..lineTo(rect.left, rect.top);
+  }
+
+  @override
+  void paint(Canvas canvas, Rect rect, {TextDirection? textDirection}) {
+    final width = rect.width;
+    final height = rect.height;
+
+    final backgroundPaint = Paint()
+      ..color = overlayColor
+      ..style = PaintingStyle.fill;
+
+    final cutOutPaint = Paint()
+      ..color = Colors.black
+      ..style = PaintingStyle.fill
+      ..blendMode = BlendMode.dstOut;
+
+    final cutOutRect = Rect.fromLTWH(
+      rect.left + width / 2 - cutOutWidth / 2,
+      rect.top + height / 2 - cutOutHeight / 2,
+      cutOutWidth,
+      cutOutHeight,
+    );
+
+    canvas
+      ..saveLayer(rect, backgroundPaint)
+      ..drawRect(rect, backgroundPaint)
+      ..drawRRect(RRect.fromRectAndRadius(cutOutRect, radius), cutOutPaint)
+      ..restore();
+  }
+
+  @override
+  ShapeBorder scale(double t) {
+    return _ScannerOverlayShape(
+      overlayColor: overlayColor,
+      cutOutWidth: cutOutWidth,
+      cutOutHeight: cutOutHeight,
+      radius: Radius.zero,
     );
   }
 }