+++ lib/src/widget/camera_widget.dart
@@ -1,23 +1,27 @@
 import 'dart:io';
+
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
+import 'package:ml_card_scanner/src/model/typedefs.dart';
+import 'package:ml_card_scanner/src/widget/camera_preview_wrapper.dart';
 
 class CameraWidget extends StatefulWidget {
   final CameraController cameraController;
   final CameraDescription cameraDescription;
   final int scannerDelay;
+  final CameraPreviewBuilder? cameraPreviewBuilder;
+  final void Function(InputImage inputImage) onImage;
 
   const CameraWidget({
     required this.cameraController,
     required this.cameraDescription,
     required this.onImage,
     required this.scannerDelay,
+    this.cameraPreviewBuilder,
     super.key,
   });
-
-  final Function(InputImage inputImage) onImage;
 
   @override
   CameraViewState createState() => CameraViewState();
@@ -48,13 +52,23 @@ class CameraViewState extends State<CameraWidget> {
 
   @override
   void initState() {
-    startCameraStream();
     super.initState();
+    startCameraStream();
   }
 
   @override
   Widget build(BuildContext context) {
-    final mediaSize = MediaQuery.of(context).size;
+    final mediaSize = MediaQuery.sizeOf(context);
+
+    if (widget.cameraPreviewBuilder != null) {
+      return widget.cameraPreviewBuilder?.call(
+        context,
+        CameraPreviewWrapper(cameraController: widget.cameraController),
+        widget.cameraController.value.previewSize,
+      ) ??
+          const SizedBox.shrink();
+    }
+
     return Transform.scale(
       scale: 1 / mediaSize.aspectRatio,
       child: Center(
@@ -79,7 +93,7 @@ class CameraViewState extends State<CameraWidget> {
       rotation = InputImageRotationValue.fromRawValue(sensorOrientation);
     } else if (Platform.isAndroid) {
       var rotationCompensation =
-          _orientations[widget.cameraController.value.deviceOrientation];
+      _orientations[widget.cameraController.value.deviceOrientation];
       if (rotationCompensation == null) return null;
       if (widget.cameraDescription.lensDirection == CameraLensDirection.front) {
         rotationCompensation = (sensorOrientation + rotationCompensation) % 360;